🛡️ Sentinel: [security improvement] Centralize cryptographically secure ID generation
Consolidate and secure unique identifier generation across the application by replacing insecure `Math.random()` and manual `crypto` fallback logic with a centralized, cryptographically secure utility.

Changes:
- Created `src/lib/security/ids.ts` for core secure ID generation logic.
- Exposed `generateSecureId` via `src/lib/security/index.ts` and `src/lib/utils.ts`.
- Refactored `Analytics`, `Logger`, `Event Bus`, `Cloudflare`, `A/B Testing`, and `Export Manager` to use the secure utility.
- Hardened session ID generation in `session-analytics.ts` and `analytics.ts`.
- Added unit tests in `tests/utils.test.ts`.
- Updated security journal in `docs/security/sentinel.md`.

This version ensures full compatibility with Cloudflare Workers by avoiding `require()` and resolving potential circular dependencies by isolating the ID generation utility.
